Add basic hub model
This commit introduces the concepts of hubs to the application. For the beginning, this only ensures there is a city and a country specified for the hub. Love Foundation defines hubs as the city name, hence there is no separate name needed.
It is defined by the city (and country if that would be needed)